+++ eCommerce.Inventory.Application/Interfaces/IAuthService.cs
@@ -5,5 +5,5 @@ namespace eCommerce.Inventory.Application.Interfaces;
 public interface IAuthService
 {
     Task<AuthResponseDto> LoginAsync(LoginDto loginDto);
-    Task<AuthResponseDto> RegisterAsync(RegisterDto registerDto);
+    Task ChangePasswordAsync(string username, ChangePasswordDto changePasswordDto);
 }